Initial glotter implementation
