+++ requirements-test.txt
@@ -2,3 +2,4 @@ docker>=4.0.2,<4.1
 Jinja2>=2.10.1,<2.11
 pytest>=5.1.2,<5.2
 PyYAML>=5.1,<5.2
+glotter==0.0.27
